make changes as per PR #6
