+++ catalog_front/src/store/modules/auth/actions.js
@@ -1,3 +1,5 @@
+import handleBadResponse from '../../../util/badResponseHandler';
+
 let timer;
 
 export default {
@@ -90,24 +92,5 @@ export default {
   autoLogout(context) {
     context.dispatch('logout');
     context.commit('setAutoLogout');
-  }
-};
-
-// Private functions
-const handleBadResponse = (response, responseData) => {
-  if (!response.ok && responseData.statusCode === 401) {
-    const error = new Error('Email/Mot de passe invalides.');
-    throw error;
-  }
-
-  if (!response.ok && responseData.statusCode === 422) {
-    const validationMessage = responseData.data.map(el => Object.values(el));
-    const error = new Error(`Erreur de validation: ${validationMessage}`);
-    throw error;
-  }
-
-  if (!response.ok && responseData.statusCode === 500) {
-    const error = new Error(responseData.message || 'Failed to signup.');
-    throw error;
   }
 };
